Add basic fps profiling
First pass at profiling fps, seems to work pretty well and is a good
place to build on, might want to design a benchmarking game to stress
test.

Would be good to be able to collect fps profile data grouped ito
sections so you can make a game that progressively gets more
performance intensive?

Could be cleaner, not super happy with the `(if profiling? ...)`
sections in core, maybe would be neater to provide a list of
plugin-like things which define their own opts and initial state maps?