Add stress test
Created a stress test game instance which will add a new animated
sprite every 100 frames and append the profiling info of the past 100
frames to a results file.

Once we implement collision detection we should update this test to
include it, or split it into a stress testing suite which tests
features in isolation.

Might be useful to ceate a bash script to compare the results of
running the stress tests on master vs the current branch to help
analyse performance characteristics of new features.