[WIP] copy sound ns from lambdahoy
This adds in the basics required to get music and sound effects
playing from my last game. There's a lot of room for improvement.

Want to make sure we can declare the sound resources upfront when the
game starts (also this should probably happen to images too).

Want to have a think about how separate sound effects and music should
be, see if there's any way of getting music to invoke a callback when
it finishes, and also there's a bug where playing too many sounds at
once crashes the sound engine.