implement rotating sprites
This was more complex than anticipated, should've tackled it earlier
in retrospect.

In order for sprites to have a rotation angle and for them to be
rotated around their center (as opposed to their top left corner) we
needed to offset all of our drawing functions and collision detection
predicates to account for the sprite being centered.

Got rid of the optional `offsets` tuple while I was doing it as it was
just becoming confusing, all sprites are now drawn centered on their
`pos`.