Improve line-line intersection algorithm.
Checking if two line segments intersect is one of the fundamental
functions when writing collision detection functions.

If you need to check if a point is inside a polygon, you do so by
tracing a ray out from the point and checking for line-line
intersections with each edge of the polygon, performing N line-line
checks for an N-sided polygon.

If you need to check if two polygons are colliding you need to check
for line intersection on each distinct pair of edges across the two
polygons, performing N*M line-line checks for an N-sided and an
M-sided polygon.

Any improvements to how fast we can calculate whether two line
segments intersect are welcome.

In our original function we calculated two ratios and declared a
intersection when both ratios were between 0 and 1 (inclusive).

The improvement we're making is not actually performing the
comparatively expensive division operation, and instead directly
comparing the numerator and denominator of the ratio, as a ratio is
only between 0 and 1 if either:

  <= abs(numerator) abs(denominator)

(assuming they both have the same sign)

At the same time this commit fixes an issue with the function
`poly-lines` which returns the lines that make up a polygon based on
its vertices. It's been updated to use the sequence abstraction as the
original function didn't work when provided a list of vertices instead
of a vector. Coincidentally its performance has increased slightly due
to the change.