Improve sound ns
We were having an issue where after playing 256 sounds the
`AudioSystem` would throw a `LineUnavailableException`, turns out we
need to clean up the Line objects when they've finished playing their
clip.

We can reify the `LineListener` interface, switching on the
`LineEvent` arg type, we can get a reference to the parent `Line`
object and call `close` as long as we do it on another thread, so we
just stick it in a future.