Let (update|draw)-scene-sprites to specify key for sprite collection
Sometimes you might want to have multiple collections of sprites on a
scene (like having a separate set to display on game-over) it's useful
to be able to update and/or draw these collections separately.

Also added useful utility function for updating sprites in the current
scene by a predicate function, as well as a utility function for
generating predicates based on sprite-groups.