Force `lighten` and `darken` to return vectors
When tweening colours we need them to be vectors so we can update an
individual rgba component at a time. The previous implementations of
lighten/darken returned lists.

Technically this is a BC breaking change since there is a real
possibility of someone conjing an alpha onto a colour sequence, and
vectors and lists have different behaviours here.