Optionally calculate and display FPS
Tracking the delta time between the last 30 frames gives us a
reasonably stable framerate which we can display if the `:display-fps`
option is set on the main state map.

Might be worth considering moving this to a separate `debug` namespace
if we add in any more features like this (thinking about stuff like
displaying bounding boxes for sprites).