Improve ability to restart games
By supplying a `restart-fn` to the game and inserting it into the
state we give ourselves the ability to request a reset form anywhere
in our game.

We avoid the previously experienced circular dependencies since we can
define our restart function in core (where our scenes are defined) and
we can invoke it while in a scene using quip rather than anything in
our own namespaces.